feat: moved everything to safir-core
