+++ safir-core/Cargo.toml
@@ -6,3 +6,12 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+colored = "2.0.4"
+dirs = "5.0.1"
+psutil = "3.2.2"
+rubin = "0.3.1"
+serde = { version = "1.0.188", features = ["derive"] }
+serde_json = "1.0.105"
+sysinfo = "0.29.9"
+tokio = { version = "1.32.0", features = ["full"] }
+which = "4.4.0"
