+++ safir-core/src/disk.rs
