+++ safir-core/src/mem.rs
