+++ safir/Cargo.toml
@@ -15,10 +15,5 @@ categories = ["command-line-utilities"]
 
 [dependencies]
 clap = { version = "4.2.5" , features = ["derive"] }
-serde = { version = "1", features = ["derive"] }
-serde_json = "1"
-dirs = "5"
-colored = "2"
-rubin = "^0.3.1"
 tokio = { version = "1.28.2", features = ["full"] }
 safir-core = { version = "0.1.0", path = "../safir-core" }
